+++ src/gym_hil/controllers/opspace.py
@@ -1,3 +1,0 @@
-"""Compatibility exports for operational-space controller."""
-
-from lerobot_sim_lab.envs.controllers.opspace import *  # noqa: F403