+++ src/gym_hil/wrappers/factory.py
@@ -1,3 +1,0 @@
-"""Compatibility exports for wrapper factory helpers."""
-
-from lerobot_sim_lab.envs.wrappers.factory import *  # noqa: F403